placed loader in component
