+++ src/app/app.module.ts
@@ -43,12 +43,11 @@ import { LogoutComponent } from '../components/logout/logout';
 import { ManageAccountModal } from '../modals/manage-account/manage-account';
 import { HomeComponent } from '../components/home/home';
 import { OrgSearchComponent } from '../components/org-search/org-search';
-import { LoadingControllerComponent } from '../components/loading-controller/loading-controller';
 
 //Modals
 import { ForWorkModal } from '../modals/for-work-modal/for-work-modal';
 import { BeAnOrganizationModal } from '../modals/be-an-organization/be-an-organization';
-
+import { LoadingControllerModal } from '../modals/loading-controller/loading-controller';
 
 
 
@@ -85,7 +84,7 @@ let injections: any[] = [
     ManageAccountModal,
     HomeComponent,
     OrgSearchComponent,
-    LoadingControllerComponent
+    LoadingControllerModal
     ]
 
 @NgModule({
