--- src/components/loading-controller/loading-controller.html
@@ -1,4 +1,0 @@
-<div>
-    <div></div>
-    <ion-spinner ></ion-spinner>
-</div>
